Refactor beacon chain to simplify block processing  - Split state.go into active_state.go and crystallized_state.go  - Convert helper methods in core.go into methods for Block, ActiveState, and CrystallizedState  - Simplify block processing logic
