+++ beacon-chain/blockchain/service.go
@@ -7,10 +7,8 @@ import (
 
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/event"
-	"github.com/prysmaticlabs/prysm/beacon-chain/casper"
 	"github.com/prysmaticlabs/prysm/beacon-chain/powchain"
 	"github.com/prysmaticlabs/prysm/beacon-chain/types"
-	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	"github.com/sirupsen/logrus"
 )
 
@@ -27,7 +25,6 @@ type ChainService struct {
 	beaconDB                       ethdb.Database
 	chain                          *BeaconChain
 	web3Service                    *powchain.Web3Service
-	isValidator                    bool
 	incomingBlockFeed              *event.Feed
 	incomingBlockChan              chan *types.Block
 	canonicalBlockFeed             *event.Feed
@@ -51,19 +48,12 @@ type Config struct {
 // be registered into a running beacon node.
 func NewChainService(ctx context.Context, cfg *Config) (*ChainService, error) {
 	ctx, cancel := context.WithCancel(ctx)
-	var isValidator bool
-	if cfg.Web3Service == nil {
-		isValidator = false
-	} else {
-		isValidator = true
-	}
 	return &ChainService{
 		ctx:                            ctx,
 		chain:                          cfg.Chain,
 		cancel:                         cancel,
 		beaconDB:                       cfg.BeaconDB,
 		web3Service:                    cfg.Web3Service,
-		isValidator:                    isValidator,
 		latestProcessedBlock:           make(chan *types.Block, cfg.BeaconBlockBuf),
 		incomingBlockChan:              make(chan *types.Block, cfg.IncomingBlockBuf),
 		incomingBlockFeed:              new(event.Feed),
@@ -77,11 +67,6 @@ func NewChainService(ctx context.Context, cfg *Config) (*ChainService, error) {
 
 // Start a blockchain service's main event loop.
 func (c *ChainService) Start() {
-	if c.isValidator {
-		log.Infof("Starting service as validator")
-	} else {
-		log.Infof("Starting service as observer")
-	}
 	// TODO: Fetch the slot: (block, state) DAGs from persistent storage
 	// to truly continue across sessions.
 	go c.blockProcessing(c.ctx.Done())
@@ -165,9 +150,8 @@ func (c *ChainService) GetCanonicalBlockBySlotNumber(slotnumber uint64) (*types.
 	return c.chain.getCanonicalBlockForSlot(slotnumber)
 }
 
-// updateHead applies the fork choice rule to the last received
-// slot.
-func (c *ChainService) updateHead(slot uint64) {
+// updateHead applies the fork choice rule to the last received slot.
+func (c *ChainService) updateHead() {
 	// Super naive fork choice rule: pick the first element at each slot
 	// level as canonical.
 	//
@@ -180,20 +164,6 @@ func (c *ChainService) updateHead(slot uint64) {
 	if err := c.chain.SetCrystallizedState(c.candidateCrystallizedState); err != nil {
 		log.Errorf("Write crystallized state to disk failed: %v", err)
 	}
-
-	// TODO: Utilize this value in the fork choice rule.
-	vals, err := casper.ShuffleValidatorsToCommittees(
-		c.candidateCrystallizedState.DynastySeed(),
-		c.candidateCrystallizedState.Validators(),
-		c.candidateCrystallizedState.CurrentDynasty(),
-		c.candidateCrystallizedState.CrosslinkingStartShard(),
-	)
-
-	if err != nil {
-		log.Errorf("Unable to get validators by slot and by shard: %v", err)
-		return
-	}
-	log.Debugf("Received %d validators by slot", len(vals))
 
 	h, err := c.candidateBlock.Hash()
 	if err != nil {
@@ -215,7 +185,8 @@ func (c *ChainService) updateHead(slot uint64) {
 	// We fire events that notify listeners of a new block (or crystallized state in
 	// the case of a state transition). This is useful for the beacon node's gRPC
 	// server to stream these events to beacon clients.
-	transition := c.chain.IsCycleTransition(slot)
+	cState := c.chain.CrystallizedState()
+	transition := cState.IsCycleTransition(c.candidateBlock.SlotNumber())
 	if transition {
 		c.canonicalCrystallizedStateFeed.Send(c.candidateCrystallizedState)
 	}
@@ -224,6 +195,43 @@ func (c *ChainService) updateHead(slot uint64) {
 	c.candidateBlock = nilBlock
 	c.candidateActiveState = nilActiveState
 	c.candidateCrystallizedState = nilCrystallizedState
+}
+
+func (c *ChainService) getParentSlot(block *types.Block) (uint64, error) {
+	parentBlock, err := c.chain.getBlock(block.ParentHash())
+	if err != nil {
+		return 0, err
+	}
+
+	return parentBlock.SlotNumber(), nil
+}
+
+func (c *ChainService) doesParentExist(block *types.Block) bool {
+	parentExists, err := c.chain.hasBlock(block.ParentHash())
+	if err != nil {
+		log.Debugf("Could not check existence of parent hash: %v", err)
+		return false
+	}
+	if !parentExists {
+		log.Debugf("parent does not exist: %x", block.ParentHash())
+		return false
+	}
+
+	return true
+}
+
+func (c *ChainService) isPoWBlockValid(block *types.Block) bool {
+	if c.web3Service == nil {
+		return true
+	}
+
+	fetcher := c.web3Service.Client()
+	if _, err := fetcher.BlockByHash(context.Background(), block.PowChainRef()); err != nil {
+		log.Debugf("fetching PoW block corresponding to mainchain reference failed: %v", err)
+		return false
+	}
+
+	return true
 }
 
 func (c *ChainService) blockProcessing(done <-chan struct{}) {
@@ -244,69 +252,21 @@ func (c *ChainService) blockProcessing(done <-chan struct{}) {
 			// Another routine will run that will continually compute
 			// the canonical block and states from this data structure using the
 			// fork choice rule.
-			var canProcessBlock bool
-			var canProcessAttestations bool
-			var err error
-			var blockVoteCache map[[32]byte]*types.VoteCache
-
-			h, err := block.Hash()
+			aState := c.chain.ActiveState()
+			cState := c.chain.CrystallizedState()
+			blockHash, err := block.Hash()
 			if err != nil {
-				log.Debugf("Could not hash incoming block: %v", err)
-			}
-
-			receivedSlotNumber := block.SlotNumber()
-
-			log.WithField("blockHash", fmt.Sprintf("0x%x", h)).Info("Received full block, processing validity conditions")
-
-			parentExists, err := c.chain.hasBlock(block.ParentHash())
-			if err != nil {
-				log.Debugf("Could not check existence of parent hash: %v", err)
-			}
-
-			// If parentHash does not exist, received block fails validity conditions.
-			if !parentExists && receivedSlotNumber > 1 {
-				continue
+				log.Debugf("Failed to get hash of block: %v", err)
+				return
 			}
 
 			// Process block as a validator if beacon node has registered, else process block as an observer.
-			canProcessBlock, err = c.chain.CanProcessBlock(c.web3Service.Client(), block, c.isValidator)
-			if err != nil {
-				// We might receive a lot of blocks that fail validity conditions,
-				// so we create a debug level log instead of an error log.
-				log.Debugf("Incoming block failed validity conditions: %v", err)
-			}
-
-			// Process attestations as a beacon chain node.
-			var processedAttestations []*pb.AttestationRecord
-			for index, attestation := range block.Attestations() {
-				if err := c.chain.processAttestation(index, block); err != nil {
-					canProcessAttestations = false
-					log.Errorf("could not process attestation for block %d because %v", block.SlotNumber(), err)
-				} else {
-					canProcessAttestations = true
-					processedAttestations = append(processedAttestations, attestation)
-				}
-			}
-
-			// If we cannot process an attestation in this block, we keep listening.
-			if !canProcessAttestations {
-				continue
-			}
-			// If we cannot process this block, we keep listening.
-			if !canProcessBlock {
+			if !c.doesParentExist(block) || !c.isPoWBlockValid(block) || !block.IsValid(aState, cState) {
 				continue
 			}
 
-			// With a valid beacon block, we can compute its attestations and store its votes/deposits in cache.
-			for index := range block.Attestations() {
-				blockVoteCache, err = c.chain.calculateBlockVoteCache(index, block)
-				if err != nil {
-					log.Debugf("could not calculate new block vote cache: %v", nil)
-				}
-			}
-
-			if c.candidateBlock != nilBlock && receivedSlotNumber > c.candidateBlock.SlotNumber() && receivedSlotNumber > 1 {
-				c.updateHead(receivedSlotNumber)
+			if c.candidateBlock != nilBlock && block.SlotNumber() > c.candidateBlock.SlotNumber() {
+				c.updateHead()
 			}
 
 			if err := c.chain.saveBlock(block); err != nil {
@@ -330,21 +290,26 @@ func (c *ChainService) blockProcessing(done <-chan struct{}) {
 			// TODO: Using latest block hash for seed, this will eventually be replaced by randao.
 
 			// Entering cycle transitions.
-			isTransition := c.chain.IsCycleTransition(receivedSlotNumber)
-			activeState := c.chain.ActiveState()
-			crystallizedState := c.chain.CrystallizedState()
-			if isTransition {
-				crystallizedState, activeState = c.chain.initCycle(crystallizedState, activeState)
+			if cState.IsCycleTransition(block.SlotNumber()) {
+				cState, err = cState.DeriveCrystallizedState(aState)
+			}
+			if err != nil {
+				log.Errorf("Failed to calculate the new fork choice rule: %v", err)
 			}
 
-			activeState, err = c.chain.computeNewActiveState(processedAttestations, activeState, blockVoteCache, h)
+			parentSlot, err := c.getParentSlot(block)
+			if err != nil {
+				log.Errorf("Failed to get parent slot of block %x", blockHash)
+				return
+			}
+			aState, err = aState.DeriveActiveState(block, cState, parentSlot)
 			if err != nil {
 				log.Errorf("Compute active state failed: %v", err)
 			}
 
 			c.candidateBlock = block
-			c.candidateActiveState = activeState
-			c.candidateCrystallizedState = crystallizedState
+			c.candidateActiveState = aState
+			c.candidateCrystallizedState = cState
 
 			log.Info("Finished processing state for candidate block")
 		}
