+++ beacon-chain/types/active_state_test.go
@@ -1,0 +1,1 @@
+package types
